chat-with-pdf/backend/requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
fastapi==0.109.2
faiss-cpu==1.7.3
ollama==0.3.2
langchain-ollama==0.1.3
uvicorn==0.24.0
python-multipart==0.0.6
pypdf==3.17.1
pydantic==2.5.2
python-dotenv==1.0.0
websockets==10.4